aarch64: Remove unnecessary casts (beta clippy check)

+++ arch/src/aarch64/fdt.rs
@@ -259,7 +259,7 @@ fn create_memory_node(
         if last_addr < super::layout::MEM_32BIT_RESERVED_START.raw_value() {
             // Case 1: all RAM is under the hole
             let mem_size = last_addr - super::layout::RAM_START.raw_value() + 1;
-            let mem_reg_prop = [super::layout::RAM_START.raw_value() as u64, mem_size as u64];
+            let mem_reg_prop = [super::layout::RAM_START.raw_value(), mem_size];
             let memory_node = fdt.begin_node("memory")?;
             fdt.property_string("device_type", "memory")?;
             fdt.property_array_u64("reg", &mem_reg_prop)?;
@@ -269,7 +269,7 @@ fn create_memory_node(
             // Region 1: RAM before the hole
             let mem_size = super::layout::MEM_32BIT_RESERVED_START.raw_value()
                 - super::layout::RAM_START.raw_value();
-            let mem_reg_prop = [super::layout::RAM_START.raw_value() as u64, mem_size as u64];
+            let mem_reg_prop = [super::layout::RAM_START.raw_value(), mem_size];
             let memory_node_name = format!("memory@{:x}", super::layout::RAM_START.raw_value());
             let memory_node = fdt.begin_node(&memory_node_name)?;
             fdt.property_string("device_type", "memory")?;
@@ -278,10 +278,7 @@ fn create_memory_node(
 
             // Region 2: RAM after the hole
             let mem_size = last_addr - super::layout::RAM_64BIT_START.raw_value() + 1;
-            let mem_reg_prop = [
-                super::layout::RAM_64BIT_START.raw_value() as u64,
-                mem_size as u64,
-            ];
+            let mem_reg_prop = [super::layout::RAM_64BIT_START.raw_value(), mem_size];
             let memory_node_name =
                 format!("memory@{:x}", super::layout::RAM_64BIT_START.raw_value());
             let memory_node = fdt.begin_node(&memory_node_name)?;
@@ -303,7 +300,7 @@ fn create_chosen_node(
     fdt.property_string("bootargs", cmdline)?;
 
     if let Some(initrd_config) = initrd {
-        let initrd_start = initrd_config.address.raw_value() as u64;
+        let initrd_start = initrd_config.address.raw_value();
         let initrd_end = initrd_config.address.raw_value() + initrd_config.size as u64;
         fdt.property_u64("linux,initrd-start", initrd_start)?;
         fdt.property_u64("linux,initrd-end", initrd_end)?;
